rename option

+++ experiments/plot.py
@@ -114,7 +114,7 @@ for idx, (name0, df0) in enumerate(df.groupby(args.group0)) if args.group0 else 
 
     if args.ref:
         rate, x, y0 = args.ref.split(';')
-        reference_triangle_loglog(int(rate), [float(xi) for xi in x.split(',')],
+        reference_triangle_loglog(float(rate), [float(xi) for xi in x.split(',')],
           float(y0), tx, ty, linewidth=1, color='.5')
 
     if args.comment: plt.text(1.0, -0.1, args.comment, transform=plt.gca().transAxes,
